New script 'update-credits.sh'
This script will update the CREDITS file's Contributors section with the
list of all people who authored commits in the current branch of the GIT
repository (using the shortlog).

The repository's .mailmap file should be updated as appropriate (see the
git shortlog man page [1] for details) prior to running this script, so
that commit made by the same person under different names and/or e-mail
addresses are properly aggregated.

[1] http://www.kernel.org/pub/software/scm/git/docs/git-shortlog.html#_mapping_authors