docbook-manual.py: allow build with publican
The script now automatically selects the appropriate toolchain to build
the docbook (make or publican), based on presence of the corresponding
config file (Makefile or publican.cfg) in the docbook manual's source
directory. If neither of these files are present, the script skips the
build.

Since publican does not include a feature to copy the built manuals to
the target location as the old makefile did, this is done manually
within the script.

Fixes #15887