New script to update Github teams access to plugin repos
The 'Core team' and 'translatewiki.net' teams in the mantisbt-plugins
org should have push access to all repositories in the organization.

Due to the large number of repositories within the org, this becomes
painful to check manually. This script leverages the Github API to
check for any missing access, and grant it.