buildrelease-repo.py: fix bug when no refs provided
When no ref to build is provided, the script retrieves the current
HEAD's SHA and uses that as release name.

The problem was that using extend() to add the SHA to the list of refs
caused the string to be added as individual chars instead of a single
item. Using append() instead fixes the problem.

Fixes #26386