Don't hardcode git commit sha length to 7
Since Git 2.11 [1], by default the abbreviated commit sha's length is
determined automatically based on the number of objects in the
repository.

As of this commit, in the mantisbt repository that gives us 9 chars,
which causes the nightly-builds.sh script's logic to keep the N most
recent builds to fail.

Changing the regex to accept any number of chars.

Fixes #28009

[1]: https://github.com/git/git/blob/master/Documentation/RelNotes/2.11.0.txt#L22